requirements.txt
SQLAlchemy==1.4.40
appdirs~=1.4.4
beautifulsoup4==4.11.1
PySide6-Essentials==6.3.1
Pillow==9.4.0
reportlab==3.6.11
packaging~=21.3
traitlets==5.3.0
requests==2.28.1
markdown2==2.4.3
lxml==4.9.1
Jinja2==3.1.2
alembic==1.8.1
aiohttp~=3.8.4